requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
python-dotenv==1.0.0
jinja2==3.1.2
pytest==7.4.3
black==23.11.0
flake8==6.1.0 